Resolved on one rule: master is the base, and only genuine time-travel work is
re-applied on top of it.

23 conflicted files. Where both sides merely added things (CLI verbs, feature
flags, notification payloads, CMake blocks, doc sections) both additions are
kept. Where the two diverged structurally, master's file is taken and the TTD
parts are re-applied onto it: emulatorcontext.h (master's layout plus
pTimeTravelManager / ttdReplayActive / ttdProbe and the run-control claim),
breakpointmanager.cpp (five silent-replay guards), tape.cpp and wd1793.cpp
(external-event markers plus the TTDSerializable implementations), covox.

Branch work that is not time-travel goes back to master: the SIMD screenzx
renderer and decimators, the AY/TurboSound rework, the handleStep batching, and
the branch's extended-key decomposition in keyboard.cpp (master already does it,
and guards base != key). The TTD serializers those files carried were ported
onto master's versions rather than dropped - that is state serialization, which
is TTD's contribution.

Three defects came from files git merged without conflict, where the text
combined but the meaning did not:

- mainloop.cpp ended up with two definitions of WaitForPauseConfirmation, one
  from each side. Master's is kept: it has the fast path for a stopped loop and
  the same-thread guard that stops a call from the emulation thread waiting on
  itself until the timeout.
- branch headers against master sources orphaned WD1793::TTDSaveState and
  ScreenZX::DrawPeriod. The FDC serializer is re-applied; DrawPeriod was the
  SIMD override, so screenzx.h comes from master too and the base implementation
  is used.
- Multirate_Test.AY_PitchInvariantAcrossRates failed by exactly the ratio of the
  core rates (176400/44100), because the branch's AY sampling path sat under
  master's new multirate design.

Also pulls the ENABLE_RECORDING ODR fix from master's tip (ca19c843). The WebAPI
translation units compiled emulatorcontext.h without that macro while core had
it, so pRecordingManager existed in one layout and not the other and every
member after it was read at the wrong offset. Upstream saw it as "Disassembler
not available"; here it was a hard crash - a single GET on /ttd/status
segfaulted in TTDCodecPageStore::GetUsedBytes, reached through a garbage
pTimeTravelManager that still passed the null check.

Verified: core, unreal-qt and core-tests build; 1582 of 1584 tests pass in
Release (2 skipped, none failed). The five BreakpointHotpathBench failures seen
in Debug are a build-type artefact - 1.3 ns/call in Release against a 5.0 ns
budget. WebAPI answers /ttd/status and both disassembly endpoints, and the TTD
scrubber connects and polls an instance.

Known, pre-existing and untouched: core/benchmarks/debugger/ttd/
ttd_frame_overhead_benchmark.cpp includes tests/_helpers/mainloop_cut.h, which
exists in no commit, so the core-benchmarks target does not build.

+++ core/tests/debugger/ttd/ttd_checkpoint_test.cpp
@@ -293,7 +293,9 @@ TEST(TTDCpuStateTest, Capture_DetectsAllArchitecturalChanges)
     EXPECT_FIELD_SEEN([](Z80State& z)->uint8_t& { return z.halted; }, 0x00);
     EXPECT_FIELD_SEEN([](Z80State& z)->uint16_t& { return z.memptr; }, 0xFFFF);
     EXPECT_FIELD_SEEN([](Z80State& z)->uint8_t& { return z.q; }, 0xFF);
-    EXPECT_FIELD_SEEN([](Z80State& z)->uint16_t& { return z.eipos; }, 0xFFFF);
+    // eipos is int32_t in the current Z80State (it stores a t-state position,
+    // not a 16-bit register), so the getter must match or the reference will not bind.
+    EXPECT_FIELD_SEEN([](Z80State& z)->int32_t& { return z.eipos; }, 0xFFFF);
     EXPECT_FIELD_SEEN([](Z80State& z)->uint16_t& { return z.haltpos; }, 0xFFFF);
     EXPECT_FIELD_SEEN([](Z80State& z)->bool& { return z.nmi_in_progress; }, false);
     EXPECT_FIELD_SEEN([](Z80State& z)->bool& { return z.int_pending; }, false);
